pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "spidershield"
version = "0.3.0"
description = "Security scanner and runtime guard for MCP servers -- static analysis, policy enforcement, DLP, prompt injection detection"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "SpiderShield" },
]
keywords = ["mcp", "ai-agent", "security", "tool-firewall", "runtime-security", "dlp", "prompt-injection", "guardrails", "static-analysis", "sarif"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Security",
    "Topic :: Software Development :: Quality Assurance",
]
dependencies = [
    "click>=8.1",
    "rich>=13.0",
    "httpx>=0.27",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "mcp>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",
]
ai = [
    "anthropic>=0.40",
    "openai>=1.0",
    "google-generativeai>=0.5",
]

[project.scripts]
spidershield = "spidershield.cli:main"
spidershield-server = "spidershield.server:run"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
